+++ packages/core/src/admin-ui/components/index.ts
@@ -1,0 +1,32 @@
+// FIELD VIEW SPECIFIC COMPONENTS
+export { CellContainer } from './CellContainer';
+export { CellLink } from './CellLink';
+
+export { ErrorBoundary, ErrorContainer } from './Errors';
+
+// ADMIN-UI CUSTOM COMPONENTS
+export { Logo } from './Logo';
+export {
+  Navigation,
+  NavigationContainer,
+  NavItem,
+  ListNavItems,
+  ListNavItem,
+  PopoverLink,
+  AuthenticatedItemDialog,
+} from './Navigation';
+
+// co-locating the type with the admin-ui/component for a more a salient mental model.
+// importing this type from @keystone-6/core/admin-ui/components is probably intuitive for a user
+export type { NavigationProps } from '../../types';
+
+// CUSTOM PAGE BUILDING UTILITIES
+export { PageContainer, Content, PageWrapper, Sidebar } from './PageContainer';
+export { CreateItemDrawer } from './CreateItemDrawer';
+export { GraphQLErrorNotice } from './GraphQLErrorNotice';
+
+// extra exports not in the admin-ui/components package: @k6js only
+export { Container } from './Container';
+export { Pagination, PaginationLabel } from './Pagination';
+export { SignoutButton } from './SignoutButton';
+export { UpdateItemsDrawer } from './UpdateItemsDrawer';